+++ src/ThreeDAnalyzer.Web/obj/Debug/net10.0/ThreeDAnalyzer.Web.AssemblyInfo.cs
@@ -13,7 +13,7 @@ using System.Reflection;
 [assembly: System.Reflection.AssemblyCompanyAttribute("ThreeDAnalyzer.Web")]
 [assembly: System.Reflection.AssemblyConfigurationAttribute("Debug")]
 [assembly: System.Reflection.AssemblyFileVersionAttribute("1.0.0.0")]
-[assembly: System.Reflection.AssemblyInformationalVersionAttribute("1.0.0+b99b7fa8a0d5900731bce3aa2210f896b3e2ca49")]
+[assembly: System.Reflection.AssemblyInformationalVersionAttribute("1.0.0+4fe6c5e5f488ec60a8aa1e4f7e406c43840e998a")]
 [assembly: System.Reflection.AssemblyProductAttribute("ThreeDAnalyzer.Web")]
 [assembly: System.Reflection.AssemblyTitleAttribute("ThreeDAnalyzer.Web")]
 [assembly: System.Reflection.AssemblyVersionAttribute("1.0.0.0")]